Move beforeBreadcrumb test to SentryOptionsTest class
