+++ sentry-kotlin-multiplatform/src/commonTest/kotlin/io/sentry/kotlin/multiplatform/BreadcrumbTest.kt
@@ -185,42 +185,4 @@ class BreadcrumbTest {
 
         breadcrumb.setData("key", "value")
     }
-
-    @Test
-    fun `Breadcrumb can be modified via callback in init and should return the modified Breadcrumb`() {
-        val options = SentryOptions()
-
-        fun mockInit(configuration: (SentryOptions) -> Unit) {
-            configuration.invoke(options)
-        }
-
-        val expectedBreadcrumb = Breadcrumb().apply {
-            message = "changed message"
-            type = "changed type"
-            category = "changed category"
-            setData(mutableMapOf("data1" to 12, "data2" to "value", "key" to "value"))
-        }
-
-        val breadcrumb = Breadcrumb().apply {
-            message = "another message"
-            type = "another type"
-            category = "another category"
-            setData(mutableMapOf("data1" to 12, "data2" to "value"))
-        }
-
-        mockInit {
-            it.beforeBreadcrumb = { breadcrumb ->
-                breadcrumb.apply {
-                    setData("key", "value")
-                    message = "changed message"
-                    type = "changed type"
-                    category = "changed category"
-                }
-            }
-        }
-
-        val modifiedBreadcrumb = breadcrumb.let(options.beforeBreadcrumb!!)
-
-        assertEquals(expectedBreadcrumb, modifiedBreadcrumb)
-    }
 }
